requirements.txt
langchain
ipykernel
python-dotenv 
langchain_community
langchain-text-splitters
langchain_google_genai
streamlit
langchain-core
langserve
fastapi
google-generativeai